// src/types/segmentCodes.ts
export enum SegmentCode {
  BUSINESS = 'BUSINESS',
  DEVELOPMENT_HOUSE = 'DEVELOPMENT_HOUSE',
  FREELANCE_DEVELOPER = 'FREELANCE_DEVELOPER',
  COMMERCE_PLATFORM = 'COMMERCE_PLATFORM',
  BILLING_PLATFORM = 'BILLING_PLATFORM',
  RETAIL_PLATFORM = 'RETAIL_PLATFORM',
  APP_COMMERCE = 'APP_COMMERCE',
  PAYMENT_ACQUIRER = 'PAYMENT_ACQUIRER',
  PAYMENT_FACILITATOR = 'PAYMENT_FACILITATOR',
  PAYMENT_GATEWAY_PROVIDER = 'PAYMENT_GATEWAY_PROVIDER',
  PAYMENT_TECHNOLOGY = 'PAYMENT_TECHNOLOGY',
  PAYMENT_TECHNOLOGY_ORCHESTRATION_PROVIDER = 'PAYMENT_TECHNOLOGY_ORCHESTRATION_PROVIDER',
  PAYMENT_ISSUER = 'PAYMENT_ISSUER',
  PAYMENT_SCHEME = 'PAYMENT_SCHEME',
  COMMERCIAL_REGULATOR = 'COMMERCIAL_REGULATOR',
  FINANCIAL_REGULATOR = 'FINANCIAL_REGULATOR',
}

export const SEGMENT_CODE_COUNT = Object.keys(SegmentCode).length;